command bags for AA per the discussion here, there's some disagreement about whether or not admin assistant should have command veteran bag options https://github.com/impstation/imp-station-14/pull/1683#discussion_r1955569839
in a suggestions thread in our discord ("Should Admin Assistant be an intern role? Or also an RP role?") it seems that people are happy for it to be both for rp and for learning, so presumably we won't add a time lockout to it, in which case i personally believe it's fine that they have these bags as options

imo it seems like an appropriate job for experienced command to take as a 'side job' when they're not really on the clock for captain or whatever.  and, frankly, i don't think these bags being an option is going to be the deciding factor as to whether command mains will play AA.  based on the suggestions thread, they'll still do it, just with basic bags

sooo we'll see, but community feedback has me thinking it's fine