fix fever dream drink upstream changed how eggs work right after i pr'd this drink
now instead of adding 5u of a 6u egg to a glass with 5u of space left, it splashes all the ingredients on the ground in tiny amounts to fit that last 1u of egg

so it makes a cocktail with raw egg floating in it, and i don't know C# to change that behavior

fuuuuck meeeee.  now the froth is thaven/carp blood.  it's fine