undershirts don't show under short sleeves resprited undershirts to not show under short sleeves
i think the original artist misinterpreted where the arms are on the sprite, or chose differently?  many times the backline of the east/west is used to represent the back, but just as many times it's used to represent the horizon of the arm

hard to put into words - see these highlighted pixels?

they're empty on sleeveless 'jumpsuit' items, which means they really should be empty on undershirts as well

it's better to lean toward the more conservative pixel placement so that short sleeve tops can be made more easily, and it still looks correct and convincing which is important.  just now it won't fuck up those tops