NPC Zombies will now see initial infected as fellow zombies.
This is a sidegrade - on one hand this allows IIs to not get killed by AI zombies (esp. now that #2805 will make Zombie AI work) and so they can remain a later threat rather than forced to pop. On the other hand it provides a way for a crewmember to potentially detect II, being that its suspicious if a Zombie ignores someone.