Added Iwirzine, a chemical that swaps the faction of the creature that ingests 15u min of it to "passive". Reacts with all chemnade components to create salt.
Added the Muldoon, a scoped tranquilizer sniper rifle,
and two chem-loaded darts; one with Iwirzine, and the other
with chloral hydrate.

Added a research that allowss the Muldoon and its darts to
be made at the security techfab. Also gave the gun and darts
to the zookeeper on round start.