Runahead pull of upstream PR 37056 - Fix NPCs stalling
Runahead pull of upstream PR 37056 - Fix NPCs stalling when too many exist.

Given that this fix isn't in #2465 and also that we keep running against this bug (limits active NPCs to 128), thought it best to take this now rather than wait for the next upstream merge after the ongoing one.