The Lesser Firearms Resprite
Updates the colours for the new upstream magazine in-hands. Adds new in-hands for Impstation magazines. Updates the colours on old Wizden ammo box sprites. New sprites for the M90-GL (now called the Valkyrie with a faster rate of fire), the Minigun (with new in-hands and wielding), the new Temperature gun (now called the RA-96 temperature gun), Blunderbuss, Hand-Made Pistol, Flintlock, Pirate Cannon, and Rocket Launcher. Updated the projectile, mussel-flash, and impact sprites for the temperature to use the assets I had made in anticipation of this gun in a previous PR.