Blaster Magnum Changes
I've been pretty unhappy with how this weapon has felt. I balanced it around the Advanced Laser Gun, which is hard to do because that's hitscan, and it meant giving it such a piss-poor rate of fire. Considering it has so much fewer shots than the advanced laser gun, I buffed its rate of fire to feel better, as well as adjusting its damage a tich to help curb that boost to DPS without sacrificing its impact. Finally, I reduced its recharge rate to make it slower than the Advanced Laser Gun to just help set it apart as well as have a bit of a give-and-take with its buff.