Some changes
Based on balance changes made by Delta-V-- more or less confirming suspicions I had but wanted to see how it all played out first. Didn't of course agree with the degree of the changes or all of the changes, but I do feel better about these stats to start with.