I added a blacklist
I figured it out and added blacklist functionality to the code. I did a blacklist instead of a whitelist because, under the umbrella of food and produce, there's a million different tags and, at the moment four tags for a blacklist is really not that bad without having to need to make any new tags at all. I can see a whitelist being useful for a "pickier eater" but I think I'd prefer to cross that bridge when we get there my brain hurts.