Echion Upkeep, again.
I was informed that you could top your glass up at a synthesizer, and that kind of interaction worries me as it would be basically an infinite source of poison, even if the poison isn't particularly strong compared to even the weaker conventional options. Also changed the sprites that show off how much is inside the synthesizer to be easier to read at a glance value.