oooaugh
im kind of really stupid and tried to fix tha merge conflict on my phone at like 2 am so of course i fucked it up badly . now that im like looking at it on a computer i can see how i fucked it so badly. this should be fixed. fuck